main 패턴수정 진행 중
