+++ repositories/postRepository.js
@@ -1,9 +1,4 @@
-const { Posts,
-    Users,
-    Comments,
-    PostsLikes,
-    PostsScraps,
-    sequelize, } = require("../models");
+const { Posts, Users, sequelize, } = require("../models");
 const { Op } = require("sequelize");
 class PostRepository {
     constructor() { }
@@ -12,7 +7,164 @@ class PostRepository {
         const targetPost = await Posts.findOne({ where: { post_id } });
         return targetPost
     }
+    getAllPost = async () => {
+        return await Posts.findAll({
+            attributes: [
+                "post_id",
+                "user_id",
+                [sequelize.col("nickname"), "nickname"],
+                [sequelize.col("user_type"), "user_type"],
+                "title",
+                "content",
+                "image",
+                "category",
+                "scrap_num",
+                "like_num",
+                "comment_num",
+                "created_at",
+                "updated_at",
+            ],
+            include: [
+                {
+                    model: Users,
+                    attributes: [],
+                },
+            ],
+            group: ["Posts.post_id"],
+            order: [["created_at", "DESC"]],
+            raw: true,
+        });
+    }
 
+    getTopPost = async ({ year, month, day, d }) => {
+        return await Posts.findAll({
+            attributes: [
+                "post_id",
+                "user_id",
+                [sequelize.col("nickname"), "nickname"],
+                [sequelize.col("user_type"), "user_type"],
+                "title",
+                "content",
+                "image",
+                "category",
+                "scrap_num",
+                "like_num",
+                "comment_num",
+                "created_at",
+                "updated_at",
+            ],
+            where: {
+                [Op.and]: [
+                    {
+                        created_at: {
+                            [Op.lt]: d,
+                        },
+                    },
+                    {
+                        created_at: {
+                            [Op.gte]: new Date(year, month, day - 6),
+                        },
+                    },
+                ],
+            },
+            include: [
+                {
+                    model: Users,
+                    attributes: [],
+                },
+            ],
+            group: ["Posts.post_id"],
+            order: [["like_num", "DESC"]],
+            raw: true,
+        });
+    }
+    getLimitPost = async ({ year, month, day, d, page }) => {
+        return await Posts.findAll({
+            attributes: [
+                "post_id",
+                "user_id",
+                [sequelize.col("nickname"), "nickname"],
+                [sequelize.col("user_type"), "user_type"],
+                "title",
+                "content",
+                "image",
+                "category",
+                "scrap_num",
+                "like_num",
+                "comment_num",
+                "created_at",
+                "updated_at",
+            ],
+            where: {
+                [Op.and]: [
+                    {
+                        created_at: {
+                            [Op.lt]: d,
+                        },
+                    },
+                    {
+                        created_at: {
+                            [Op.gte]: new Date(year, month, day - 6),
+                        },
+                    },
+                ],
+            },
+            include: [
+                {
+                    model: Users,
+                    attributes: [],
+                },
+            ],
+            group: ["Posts.post_id"],
+            order: [["like_num", "DESC"]],
+            offset: (page - 1) * 10,
+            limit: 10,
+            raw: true,
+        })
+    }
+    getPostByKeyword = async ({ keyword, page }) => {
+        return await Posts.findAll({
+            attributes: [
+                "post_id",
+                "user_id",
+                [sequelize.col("nickname"), "nickname"],
+                [sequelize.col("user_type"), "user_type"],
+                "title",
+                "content",
+                "image",
+                "category",
+                "scrap_num",
+                "like_num",
+                "comment_num",
+                "created_at",
+                "updated_at",
+            ],
+            include: [
+                {
+                    model: Users,
+                    attributes: [],
+                },
+            ],
+            where: {
+                [Op.or]: [
+                    {
+                        title: {
+                            [Op.like]: "%" + keyword + "%",
+                        },
+                    },
+                    {
+                        content: {
+                            [Op.like]: "%" + keyword + "%",
+                        },
+                    },
+                ],
+            },
+            order: [["created_at", "DESC"]],
+            offset: (page - 1) * 10,
+            limit: 10,
+            raw: true,
+        });
+    }
     createPost = async ({
         user_id,
         nickname,
@@ -97,7 +249,7 @@ class PostRepository {
         )
     }
 
-    updateLike = async({ like_num, post_id }) => {
+    updateLike = async ({ like_num, post_id }) => {
         await Posts.update(
             {
                 like_num: like_num.length,
@@ -108,7 +260,7 @@ class PostRepository {
         );
     }
 
-    updateScrap = async({ scrap_num, post_id }) => {
+    updateScrap = async ({ scrap_num, post_id }) => {
         await Posts.update(
             {
                 scrap_num: scrap_num.length,
