+++ repositories/postScrapRepository.js
@@ -10,6 +10,14 @@ class PostScrapRepository {
         });
     }
 
+    findByUserId = async({user_id,item}) => {
+        return await PostsScraps.findOne({
+            attributes: ["post_id", "user_id"],
+            where: { user_id: user_id, post_id: item.post_id },
+            raw: true,
+        })
+    }
+
     deleteScrap = async({user_id,post_id}) => {
         await PostsScraps.destroy({
             where: {
